Fixed up a hard dependency FormBundleWidget had on SettingWindow specific code.
